refactor: the animation's per-frame state is one struct
Twelve field names were written out three times: the list
anim_state_reset clears, the list placement() copies into CPlacement, and
the list CPlacement::geometry() copies into Placement. Nothing tied them
together, so a field added to the decoder had three places to be
remembered and no gate that noticed if it was not.

AnimState is the twelve. The decoder holds one, so the reset is an
assignment; Placement is a canvas plus one; and CPlacement holds a
Placement, so geometry() is gone and the compositor takes &pl.geom.
key_frame stays the field placement() cannot fill -- is_key_frame decides
it from the rest -- which is now the one field the struct update spells
out.

anmf_flags and prev_anmf_flags become u8, which is what the file carries
and what Placement already held; that drops four casts, including one
that widened ANMF_FLAG_DISPOSE to compare against it.

frame_index moves in with them. It was declared beside the still-image
fields but reset and read as part of the animation, which is what
grouping the rest made visible.

md5check, clicheck and animcheck are unchanged, and animation decode does
not move.

+++ src/driver/export.rs
@@ -43,7 +43,7 @@ pub struct ExportSettings {
     pub duration: i32,
     pub pos_x: i32,
     pub pos_y: i32,
-    pub anmf_flags: i32,
+    pub anmf_flags: u8,
     pub has_alpha: bool,
     pub timestamp: i64,
 }
@@ -85,7 +85,7 @@ pub(crate) fn export_frame(
     format: i32,
     out: &mut Handout<'_>,
 ) {
-    let flags = set.anmf_flags as u8;
+    let flags = set.anmf_flags;
 
     out.pixels = Pixels::None;
     out.format = Format::from_raw(format).unwrap_or(Format::Argb);
